Updatated project structure
