+++ todoPrezesy/requirements.txt
@@ -1,0 +1,3 @@
+discord.py
+python-dotenv
+pytest
